custom hook for filtering was added
